Seokmin2

See merge request sw_track/class_05/web_project/team11/webproject!61

+++ views/adminpage_order/adminpage_order.js
@@ -2,31 +2,35 @@ import fetchData from "./order_getdata.js";
 import deleteData from "./order_deletedata.js";
 import ModifyData from "./order_Modify.js";
 document.addEventListener("DOMContentLoaded", async function () {
-  try{
-        //주문 조회
-        const result = await fetchData();
-        console.log(result);
-        //주문 수정
-        const selectElement = document.querySelectorAll("#shipping");
-        console.log(selectElement);
-        selectElement.forEach((select, index) => {
-          select.addEventListener("change", async function(event) {
-            const status = selectElement[index].value;
-             await ModifyData(status); 
-          });
-        });
-        //삭제 기능
-        const deleteBtn = document.querySelectorAll(".delete-btn");
-        deleteBtn.forEach(button => {
-        button.addEventListener("click", async function (event) {
-        const orderId = event.target.dataset.orderid;
-        console.log(orderId);
-        const trElement = this.parentNode.parentNode;
-        trElement.remove();
-        await deleteData(orderId);
-        });
+  try {
+    //주문 조회
+    const result = await fetchData();
+    //주문 수정
+    const selectElement = document.querySelectorAll("#shipping");
+    selectElement.forEach((select, index) => {
+      select.addEventListener("change", async function (event) {
+        const status = selectElement[index].value;
+        await ModifyData(status);
+      });
     });
-  }catch(err){
+    //삭제 기능
+    const deleteBtn = document.querySelectorAll(".delete-btn");
+    deleteBtn.forEach((button, i) => {
+      button.addEventListener("click", async function (event) {
+        const orderName = result.orders[i].name;
+        const confirmDelete = confirm(
+          `${orderName}님 주문을 삭제 하시겠습니까??`
+        );
+        if (confirmDelete) {
+          const orderId = event.target.dataset.orderid;
+          console.log(orderId);
+          const trElement = this.parentNode.parentNode;
+          trElement.remove();
+          await deleteData(orderId);
+        }
+      });
+    });
+  } catch (err) {
     console.log(err);
   }
 });
